refactor: separate out sorting utility functions
